refactor(cli): move `Input` to its own module
